# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Workcraft API
  description: |
    Studio-ops API for Workcraft. Response envelope: `{ status: "success"|"error", data? | message? }`.
    Authenticated routes expect `Authorization: Bearer <JWT>`. Domain routes require studio membership on the token.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          example: error
        message:
          type: string
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: success
        data:
          description: Endpoint-specific payload
...
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login success — accessToken, user, studio, access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user, studio, and access
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
  /api/settings/studio:
    get:
      tags: [Settings]
      summary: Get studio settings
      operationId: getStudioSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Studio row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Settings]
      summary: Patch studio settings
      operationId: patchStudioSettings
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated studio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/options:
    get:
      tags: [Settings]
      summary: List option catalogs
      operationId: listOptionCatalogs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Catalogs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/options/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          enum:
            [
              CEREMONY_TYPE,
              JOB_TYPE,
              LEAD_SOURCE,
              PAY_METHOD,
              EXPENSE_CAT,
              INCOME_CAT,
              CANCEL_REASON,
              BANK,
              CREW_ROLE,
              CRAFT,
              EQUIP_LOCATION,
            ]
    get:
      tags: [Settings]
      summary: Get one option catalog
      operationId: getOptionCatalog
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/options/{key}/seed:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Settings]
      summary: Seed built-in option catalog
      operationId: seedOptionCatalog
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Seeded catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/options/{key}/items:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Settings]
      summary: Add option item
      operationId: addOptionItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/settings/options/items/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    patch:
      tags: [Settings]
      summary: Update or retire option item
      operationId: updateOptionItem
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                sortOrder:
                  type: integer
                retire:
                  type: boolean
      responses:
        "200":
          description: Updated catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
  /api/team/members:
    get:
      tags: [Team]
      summary: List team members
      operationId: listTeamMembers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Team]
      summary: Create team login (ADMIN/MEMBER)
      operationId: createTeamMember
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, access]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  minLength: 8
                phone:
                  type: string
                  nullable: true
                access:
                  type: string
                  enum: [ADMIN, MEMBER]
                craftItemIds:
                  type: array
                  items:
                    type: integer
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/team/members/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Team]
      summary: Get team member
      operationId: getTeamMember
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Team]
      summary: Update team member
      operationId: updateTeamMember
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Team]
      summary: Remove team member
      operationId: deleteTeamMember
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...